improved handling of negative ArrayBuffer size fixes #708

+++ src/org/mozilla/javascript/typedarrays/NativeFloat32Array.java
@@ -36,7 +36,7 @@ public class NativeFloat32Array
 
     public NativeFloat32Array(int len)
     {
-        this(new NativeArrayBuffer(len * BYTES_PER_ELEMENT), 0, len);
+        this(new NativeArrayBuffer((double)len * BYTES_PER_ELEMENT), 0, len);
     }
 
     @Override
